+++ packages/doc-gen/src/__tests__/fixtures/NestedComponent.tsx
@@ -1,0 +1,18 @@
+import React from 'react'
+
+function SubComponent({foo}: {foo: string}) {
+  return <div>{foo}</div>
+}
+
+function SubComponent2({baz}: {baz: number}) {
+  return <div>{baz}</div>
+}
+
+function RootComponent({bar}: {bar: string}) {
+  return <div>{bar}</div>
+}
+
+export const NestedComponent = Object.assign(RootComponent, {
+  SubComponent,
+  SubComponent2,
+})