+++ packages/styled-react/src/deprecated.tsx
@@ -1,1 +1,4 @@
-export {Dialog, Octicon, TabNav, Tooltip} from '@primer/react/deprecated'
+export {TabNav, type TabNavProps, type TabNavLinkProps} from './components/TabNav'
+export {Dialog, type DialogProps, type DialogHeaderProps} from './components/DialogV1'
+export {Octicon, type OcticonProps} from './components/Octicon'
+export {DeprecatedTooltip as Tooltip, type DeprecatedTooltipProps as TooltipProps} from './components/Tooltip'